refactor(webxrrenderwindowhelper): refactor WebXR rendering support
Refactors WebXR rendering support. Under previous behavior WebXR rendering was embedded in
vtkOpenGLRenderWindow. This change removes the dependency of vtkOpenGLRenderWindow on WebXR and
moves XR components into vtkWebXRRenderWindowHelper. The helper class holds a reference to an
underlying render window and is responsible to start/run/stop XR rendering.

BREAKING CHANGE: Removes WebXR API from `vtkOpenGLRenderWindow`. Applications requiring XR support
must instantiate `vtkWebXRRenderWindowHelper` instead.
