+++ Sources/Rendering/OpenGL/RenderWindow/index.d.ts
@@ -24,10 +24,6 @@ export interface IOpenGLRenderWindowInitialValues {
 	webgl2?: boolean;
 	defaultToWebgl2?: boolean;
 	activeFramebuffer?: any;
-	xrSession?: any;
-	xrSessionIsAR?: boolean;
-	xrReferenceSpace?: any;
-	xrSupported?: boolean;
 	imageFormat?: 'image/png';
 	useOffScreen?: boolean;
 	useBackgroundImage?: boolean;
@@ -230,36 +226,6 @@ export interface vtkOpenGLRenderWindow extends vtkOpenGLRenderWindowBase {
 	 * @param {I3DContextOptions} options
 	 */
 	get3DContext(options: I3DContextOptions): Nullable<WebGLRenderingContext>;
-
-	/**
-	 * Request an XR session on the user device with WebXR,
-   * typically in response to a user request such as a button press.
-	 */
-	startXR(): void;
-
-  /**
-   * When an XR session is available, set up the XRWebGLLayer
-   * and request the first animation frame for the device
-   */
-   enterXR(): void,
-
-   /**
-    * Adjust world-to-physical parameters for different viewing modalities
-    *
-    * @param {Number} inputRescaleFactor
-    * @param {Number} inputTranslateZ
-    */
-   resetXRScene(inputRescaleFactor: number, inputTranslateZ: number): void,
-
-	/**
-	 * Request to stop the current XR session
-	 */
-	stopXR(): void;
-
-	/**
-	 *
-	 */
-	xrRender(): void;
 
 	/**
 	 *
