+++ Sources/Rendering/OpenGL/RenderWindow/index.js
@@ -8,14 +8,9 @@ import vtkOpenGLTextureUnitManager from 'vtk.js/Sources/Rendering/OpenGL/Texture
 import vtkOpenGLViewNodeFactory from 'vtk.js/Sources/Rendering/OpenGL/ViewNodeFactory';
 import vtkRenderPass from 'vtk.js/Sources/Rendering/SceneGraph/RenderPass';
 import vtkRenderWindowViewNode from 'vtk.js/Sources/Rendering/SceneGraph/RenderWindowViewNode';
-import Constants from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow/Constants';
-import {
-  createContextProxyHandler,
-  GET_UNDERLYING_CONTEXT,
-} from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow/ContextProxy';
+import { createContextProxyHandler } from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow/ContextProxy';
 
 const { vtkDebugMacro, vtkErrorMacro } = macro;
-const { XrSessionTypes } = Constants;
 
 const SCREENSHOT_PLACEHOLDER = {
   position: 'absolute',
@@ -23,11 +18,6 @@ const SCREENSHOT_PLACEHOLDER = {
   left: 0,
   width: '100%',
   height: '100%',
-};
-
-const DEFAULT_RESET_FACTORS = {
-  rescaleFactor: 0.25, // isotropic scale factor reduces apparent size of objects
-  translateZ: -1.5, // default translation initializes object in front of camera
 };
 
 function checkRenderTargetSupport(gl, format, type) {
@@ -275,223 +265,6 @@ function vtkOpenGLRenderWindow(publicAPI, model) {
     }
 
     return new Proxy(result, cachingContextHandler);
-  };
-
-  // Request an XR session on the user device with WebXR,
-  // typically in response to a user request such as a button press
-  publicAPI.startXR = (xrSessionType) => {
-    if (navigator.xr === undefined) {
-      throw new Error('WebXR is not available');
-    }
-
-    model.xrSessionType =
-      xrSessionType !== undefined ? xrSessionType : XrSessionTypes.HmdVR;
-    const isXrSessionAR = [
-      XrSessionTypes.HmdAR,
-      XrSessionTypes.MobileAR,
-    ].includes(model.xrSessionType);
-    const sessionType = isXrSessionAR ? 'immersive-ar' : 'immersive-vr';
-    if (!navigator.xr.isSessionSupported(sessionType)) {
-      if (isXrSessionAR) {
-        throw new Error('Device does not support AR session');
-      } else {
-        throw new Error('VR display is not available');
-      }
-    }
-    if (model.xrSession === null) {
-      navigator.xr.requestSession(sessionType).then(publicAPI.enterXR, () => {
-        throw new Error('Failed to create XR session!');
-      });
-    } else {
-      throw new Error('XR Session already exists!');
-    }
-  };
-
-  // When an XR session is available, set up the XRWebGLLayer
-  // and request the first animation frame for the device
-  publicAPI.enterXR = async (xrSession) => {
-    model.xrSession = xrSession;
-    model.oldCanvasSize = model.size.slice();
-
-    if (model.xrSession !== null) {
-      const gl = publicAPI.get3DContext();
-      await gl.makeXRCompatible();
-
-      const glLayer = new global.XRWebGLLayer(
-        model.xrSession,
-        // constructor needs unproxied context
-        gl[GET_UNDERLYING_CONTEXT]()
-      );
-      publicAPI.setSize(glLayer.framebufferWidth, glLayer.framebufferHeight);
-
-      model.xrSession.updateRenderState({
-        baseLayer: glLayer,
-      });
-
-      model.xrSession.requestReferenceSpace('local').then((refSpace) => {
-        model.xrReferenceSpace = refSpace;
-      });
-
-      // Initialize transparent background for augmented reality session
-      const isXrSessionAR = [
-        XrSessionTypes.HmdAR,
-        XrSessionTypes.MobileAR,
-      ].includes(model.xrSessionType);
-      if (isXrSessionAR) {
-        const ren = model.renderable.getRenderers()[0];
-        model.preXrSessionBackground = ren.getBackground();
-        ren.setBackground([0, 0, 0, 0]);
-      }
-
-      publicAPI.resetXRScene();
-
-      model.renderable.getInteractor().switchToXRAnimation();
-      model.xrSceneFrame = model.xrSession.requestAnimationFrame(
-        publicAPI.xrRender
-      );
-    } else {
-      throw new Error('Failed to enter XR with a null xrSession.');
-    }
-  };
-
-  publicAPI.resetXRScene = (
-    rescaleFactor = DEFAULT_RESET_FACTORS.rescaleFactor,
-    translateZ = DEFAULT_RESET_FACTORS.translateZ
-  ) => {
-    // Adjust world-to-physical parameters for different modalities
-
-    const ren = model.renderable.getRenderers()[0];
-    ren.resetCamera();
-
-    const camera = ren.getActiveCamera();
-    let physicalScale = camera.getPhysicalScale();
-    const physicalTranslation = camera.getPhysicalTranslation();
-
-    const rescaledTranslateZ = translateZ * physicalScale;
-    physicalScale /= rescaleFactor;
-    physicalTranslation[2] += rescaledTranslateZ;
-
-    camera.setPhysicalScale(physicalScale);
-    camera.setPhysicalTranslation(physicalTranslation);
-    // Clip at 0.1m, 100.0m in physical space by default
-    camera.setClippingRange(0.1 * physicalScale, 100.0 * physicalScale);
-  };
-
-  publicAPI.stopXR = async () => {
-    if (navigator.xr === undefined) {
-      // WebXR polyfill not available so nothing to do
-      return;
-    }
-
-    if (model.xrSession !== null) {
-      model.xrSession.cancelAnimationFrame(model.xrSceneFrame);
-      model.renderable.getInteractor().returnFromXRAnimation();
-      const gl = publicAPI.get3DContext();
-      gl.bindFramebuffer(gl.FRAMEBUFFER, null);
-
-      await model.xrSession.end().catch((error) => {
-        if (!(error instanceof DOMException)) {
-          throw error;
-        }
-      });
-      model.xrSession = null;
-    }
-
-    if (model.oldCanvasSize !== undefined) {
-      publicAPI.setSize(...model.oldCanvasSize);
-    }
-
-    // Reset to default canvas
-    const ren = model.renderable.getRenderers()[0];
-
-    if (model.preXrSessionBackground != null) {
-      ren.setBackground(model.preXrSessionBackground);
-      model.preXrSessionBackground = null;
-    }
-
-    ren.getActiveCamera().setProjectionMatrix(null);
-    ren.resetCamera();
-
-    ren.setViewport(0.0, 0, 1.0, 1.0);
-    publicAPI.traverseAllPasses();
-  };
-
-  publicAPI.xrRender = async (t, frame) => {
-    const xrSession = frame.session;
-    const isXrSessionHMD = [
-      XrSessionTypes.HmdVR,
-      XrSessionTypes.HmdAR,
-    ].includes(model.xrSessionType);
-
-    model.renderable
-      .getInteractor()
-      .updateXRGamepads(xrSession, frame, model.xrReferenceSpace);
-
-    model.xrSceneFrame = model.xrSession.requestAnimationFrame(
-      publicAPI.xrRender
-    );
-
-    const xrPose = frame.getViewerPose(model.xrReferenceSpace);
-
-    if (xrPose) {
-      const gl = publicAPI.get3DContext();
-
-      if (
-        model.xrSessionType === XrSessionTypes.MobileAR &&
-        model.oldCanvasSize !== undefined
-      ) {
-        gl.canvas.width = model.oldCanvasSize[0];
-        gl.canvas.height = model.oldCanvasSize[1];
-      }
-
-      const glLayer = xrSession.renderState.baseLayer;
-      gl.bindFramebuffer(gl.FRAMEBUFFER, glLayer.framebuffer);
-      gl.clear(gl.COLOR_BUFFER_BIT);
-      gl.clear(gl.DEPTH_BUFFER_BIT);
-      publicAPI.setSize(glLayer.framebufferWidth, glLayer.framebufferHeight);
-
-      // get the first renderer
-      const ren = model.renderable.getRenderers()[0];
-
-      // Do a render pass for each eye
-      xrPose.views.forEach((view, index) => {
-        const viewport = glLayer.getViewport(view);
-
-        if (isXrSessionHMD) {
-          if (view.eye === 'left') {
-            ren.setViewport(0, 0, 0.5, 1.0);
-          } else if (view.eye === 'right') {
-            ren.setViewport(0.5, 0, 1.0, 1.0);
-          } else {
-            // No handling for non-eye viewport
-            return;
-          }
-        } else if (model.xrSessionType === XrSessionTypes.LookingGlassVR) {
-          const startX = viewport.x / glLayer.framebufferWidth;
-          const startY = viewport.y / glLayer.framebufferHeight;
-          const endX = (viewport.x + viewport.width) / glLayer.framebufferWidth;
-          const endY =
-            (viewport.y + viewport.height) / glLayer.framebufferHeight;
-          ren.setViewport(startX, startY, endX, endY);
-        } else {
-          ren.setViewport(0, 0, 1, 1);
-        }
-
-        ren
-          .getActiveCamera()
-          .computeViewParametersFromPhysicalMatrix(
-            view.transform.inverse.matrix
-          );
-        ren.getActiveCamera().setProjectionMatrix(view.projectionMatrix);
-
-        publicAPI.traverseAllPasses();
-      });
-
-      // Reset scissorbox before any subsequent rendering to external displays
-      // on frame end, such as rendering to a Looking Glass display.
-      gl.scissor(0, 0, glLayer.framebufferWidth, glLayer.framebufferHeight);
-      gl.disable(gl.SCISSOR_TEST);
-    }
   };
 
   publicAPI.restoreContext = () => {
@@ -1308,9 +1081,6 @@ const DEFAULT_VALUES = {
   webgl2: false,
   defaultToWebgl2: true, // attempt webgl2 on by default
   activeFramebuffer: null,
-  xrSession: null,
-  xrReferenceSpace: null,
-  xrSupported: true,
   imageFormat: 'image/png',
   useOffScreen: false,
   useBackgroundImage: false,
@@ -1363,9 +1133,7 @@ export function extend(publicAPI, model, initialValues = {}) {
     'shaderCache',
     'textureUnitManager',
     'webgl2',
-    'vrDisplay',
     'useBackgroundImage',
-    'xrSupported',
     'activeFramebuffer',
   ]);
